Extend SQL query row: arbitrary select columns + avg()/sum()
Two grammar extensions:
- select Col1, Col2, ... from stream [where ...] - a specific column
  projection (not the full event shape), new Table result kind.
- avg(SeverityNumber) / sum(SeverityNumber) alongside count(*) -
  SeverityNumber is the only numeric column exposed today; avg/sum on
  any other column is a parse error with a clear message. group by
  now accepts any aggregate select, not just count(*).

LogQlColumn gains SeverityNumber, but only for select/aggregate - the
parser explicitly rejects it in a where comparison (every literal
there is a string, so a numeric column would silently type-mismatch).

LogQlSelectKind (Count/Raw) replaced with a small LogQlSelect
hierarchy (Star/Columns/Aggregate) to represent the wider grammar.

Backend plumbing:
- LogAggregateQueryBuilder.BuildFromFilterSql takes an optional
  aggregate SQL fragment (default "count()", unchanged for every
  existing /api/logs/aggregate caller) so the SQL-query-row path can
  swap in avg()/sum() while reusing the same bucketing SQL.
- LogAggregateBucket.Count and LogQlQueryResponse.Count widen from
  long to double (avg() is fractional) - JSON-transparent for the
  existing whole-number count() case, no frontend type change needed.
- New Table dispatch/response kind: selected columns' real names for
  SQL, display names + generic stringified rows for the response.
- LogQl's own Count/Series aggregate SQL is always toFloat64(...)-
  wrapped so the reader is one uniform GetDouble(...) regardless of
  which function ran, with a NaN guard (avg() over zero matching rows
  is 0/0) so an otherwise-valid "nothing matched" query doesn't 500.

Frontend: SqlQueryRow.svelte renders the new Table kind as a compact
dynamic-column table; sql-highlight.ts colors avg/sum and
SeverityNumber the same as the rest of the grammar.

Verified: dotnet build clean; 505/505 Flare.Api.Tests pass (10 new,
covering both extensions + the group-by/where guard rails);
svelte-check clean. No live e2e this round per user's steer on token
usage for incremental follow-ups.

+++ src/Flare.Api/Model/LogAggregateRequest.cs
@@ -20,14 +20,21 @@ public sealed record LogAggregateRequest
     public LogAggregateGroupBy GroupBy { get; init; } = LogAggregateGroupBy.None;
 }
 
-/// <summary>One bucketed count. <see cref="GroupKey"/> is null when <see cref="LogAggregateGroupBy.None"/> was requested.</summary>
+/// <summary>
+/// One bucketed value. <see cref="GroupKey"/> is null when <see cref="LogAggregateGroupBy.None"/>
+/// was requested. <see cref="Count"/> is <c>double</c> (not <c>long</c>) so the same shape
+/// can carry a SQL-query-row <c>avg()</c>/<c>sum()</c> result (see
+/// <c>Query.LogQl.LogQlQueryBuilder</c>) as well as this endpoint's own always-integral
+/// <c>count()</c> - a whole-number value still round-trips through JSON exactly (e.g. `25`,
+/// not `25.0`), so this is a no-op for every existing <c>/api/logs/aggregate</c> caller.
+/// </summary>
 public sealed record LogAggregateBucket
 {
     public required DateTimeOffset BucketStart { get; init; }
 
     public string? GroupKey { get; init; }
 
-    public required long Count { get; init; }
+    public required double Count { get; init; }
 }
 
 /// <summary>Response body for <c>POST /api/logs/aggregate</c>.</summary>
